requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
python-jose==3.3.0
pydantic
httpx
passlib
python-multipart